src/requirements.txt
aiohttp==3.10.10
coloredlogs==15.0.1
uvloop==0.21.0
aiofiles==24.1.0
pillow
numpy
opencv-python